requirements-book.txt
# Jupyter Book and JupyterLite requirements
jupyter-book>=0.15.0
jupyterlite>=0.1.0
jupyterlite-pyodide-kernel>=0.0.7

# Core workshop dependencies
requests>=2.28.0
pandas>=1.5.0
plotly>=5.11.0
ipywidgets>=8.0.0
python-dotenv>=0.19.0

# Additional utilities
websocket-client>=1.4.0
numpy>=1.21.0
json5>=0.9.0
